Fixes #18034: split session-miner extractors to clear Qlty smells
